fix 3

+++ Sanity/rules-d/runtest.sh
@@ -81,11 +81,12 @@ rules_d_resolve_older_fapolicyd_nvr() {
   [[ -z $inst_epoch || $inst_epoch == '(none)' ]] && inst_epoch=0
   inst_evr="${inst_epoch}:${V}-${R}"
 
-  rlRun -s "dnf -q repoquery --enablerepo='*' --available --latest-limit=1 --qf '%{epoch} %{version} %{release}' \"fapolicyd < ${inst_evr}\"" 0 "Resolve latest older fapolicyd NVR"
+  rlRun -s "dnf -q repoquery --enablerepo='*' --available --show-duplicates --latest-limit=1 --qf '%{epoch} %{version} %{release}' \"fapolicyd < ${inst_evr}\"" 0 "Resolve latest older fapolicyd NVR"
   nvr_line=$(awk 'NF == 3 && $1 !~ /:$/ { print; exit }' "$rlRun_LOG")
   IFS=' ' read -r cand_epoch V_old R_old <<<"$nvr_line"
   if [[ -z ${V_old:-} || -z ${R_old:-} ]]; then
     rlLogError "no fapolicyd in repos older than installed ${V}-${R} (EVR ${inst_evr})"
+    rlLogInfo "repoquery output: $(tr '\n' ' ' < "$rlRun_LOG")"
     return 1
   fi
   [[ -z $cand_epoch || $cand_epoch == '(none)' ]] && cand_epoch=0
@@ -98,6 +99,7 @@ rules_d_resolve_older_fapolicyd_nvr() {
 }
 
 PACKAGE="fapolicyd"
+HAVE_OLD_NVR=0
 rlJournalStart && {
   rlPhaseStartSetup && {
     rlRun "rlImport --all" 0 "Import libraries" || rlDie "cannot continue"
@@ -126,16 +128,20 @@ rlJournalStart && {
     pushd rpms
     rlRun "cp $(grep 'Wrote:' $rlRun_LOG | cut -d ' ' -f 2 | tr '\n' ' ') $(grep 'Wrote:' $rlRun_LOG1 | cut -d ' ' -f 2 | tr '\n' ' ') ./"
     packages=()
-    rules_d_resolve_older_fapolicyd_nvr || rlDie "cannot resolve older fapolicyd NVR from repos"
-    if [[ -n $(dnf repoquery --enablerepo='*' --available -q "fapolicyd-dnf-plugin = ${V_old}-${R_old}" 2>/dev/null) ]]; then
-      packages+=(fapolicyd-dnf-plugin-${V_old}-${R_old}.noarch)
+    if rules_d_resolve_older_fapolicyd_nvr; then
+      HAVE_OLD_NVR=1
+      if [[ -n $(dnf repoquery --enablerepo='*' --available -q "fapolicyd-dnf-plugin = ${V_old}-${R_old}" 2>/dev/null) ]]; then
+        packages+=(fapolicyd-dnf-plugin-${V_old}-${R_old}.noarch)
+      fi
+      packages+=(
+        fapolicyd-${V_old}-${R_old}.$A
+        #fapolicyd-debuginfo-${V_old}-${R_old}.$A
+        #fapolicyd-debugsource-${V_old}-${R_old}.$A
+        fapolicyd-selinux-${V_old}-${R_old}.noarch
+      )
+    else
+      rlLogWarning "Older fapolicyd NVR not available in current repos; tests requiring upgrade from old version will be skipped"
     fi
-    packages+=(
-      fapolicyd-${V_old}-${R_old}.$A
-      #fapolicyd-debuginfo-${V_old}-${R_old}.$A
-      #fapolicyd-debugsource-${V_old}-${R_old}.$A
-      fapolicyd-selinux-${V_old}-${R_old}.noarch
-    )
 
     for package in "${packages[@]}"; do
       rlRpmDownload $package
@@ -211,31 +217,39 @@ EOF
     rlPhaseEnd; }
 
     rlPhaseStartTest "upgrade from old version - default rules" && {
-      # fapolicyd.rules should be replace with populated rules.d
-      rlRun "rm -rf /etc/fapolicyd"
-      rlRun "dnf install fapolicyd-$V_old-$R_old -y --allowerasing"
-      rlRun "dnf reinstall fapolicyd-$V_old-$R_old -y --allowerasing"
-      rlRun "ls -la /etc/fapolicyd/"
-      rlRun "dnf install fapolicyd-$V-$R -y --allowerasing"
-      rlRun "ls -la /etc/fapolicyd/"
-      rlRun "ls -la /etc/fapolicyd/rules.d/"
-      rlAssertNotExists /etc/fapolicyd/fapolicyd.rules
-      rlAssertGreater "rules are deployed into /etc/fapolicyd/rules.d" $(ls -1 /etc/fapolicyd/rules.d | wc -w) 0
+      if (( HAVE_OLD_NVR == 0 )); then
+        rlLogWarning "Skipping: older fapolicyd NVR unavailable in enabled repos"
+      else
+        # fapolicyd.rules should be replace with populated rules.d
+        rlRun "rm -rf /etc/fapolicyd"
+        rlRun "dnf install fapolicyd-$V_old-$R_old -y --allowerasing"
+        rlRun "dnf reinstall fapolicyd-$V_old-$R_old -y --allowerasing"
+        rlRun "ls -la /etc/fapolicyd/"
+        rlRun "dnf install fapolicyd-$V-$R -y --allowerasing"
+        rlRun "ls -la /etc/fapolicyd/"
+        rlRun "ls -la /etc/fapolicyd/rules.d/"
+        rlAssertNotExists /etc/fapolicyd/fapolicyd.rules
+        rlAssertGreater "rules are deployed into /etc/fapolicyd/rules.d" $(ls -1 /etc/fapolicyd/rules.d | wc -w) 0
+      fi
     rlPhaseEnd; }
 
     rlPhaseStartTest "upgrade from old version - changed rules" && {
-      # fapolicyd.rules should stay untouched
-      # rules.d should not be populated
-      rlRun "rm -rf /etc/fapolicyd"
-      rlRun "dnf install fapolicyd-$V_old-$R_old -y --allowerasing"
-      rlRun "dnf reinstall fapolicyd-$V_old-$R_old -y --allowerasing"
-      echo "allow perm=any all : all" >> /etc/fapolicyd/fapolicyd.rules
-      rlRun "ls -la /etc/fapolicyd/"
-      rlRun "dnf install fapolicyd-$V-$R -y --allowerasing"
-      rlRun "ls -la /etc/fapolicyd/"
-      rlRun "ls -la /etc/fapolicyd/rules.d/"
-      rlAssertExists /etc/fapolicyd/fapolicyd.rules
-      rlAssertEquals "rules are deployed into /etc/fapolicyd/rules.d" $(ls -1 /etc/fapolicyd/rules.d | wc -w) 0
+      if (( HAVE_OLD_NVR == 0 )); then
+        rlLogWarning "Skipping: older fapolicyd NVR unavailable in enabled repos"
+      else
+        # fapolicyd.rules should stay untouched
+        # rules.d should not be populated
+        rlRun "rm -rf /etc/fapolicyd"
+        rlRun "dnf install fapolicyd-$V_old-$R_old -y --allowerasing"
+        rlRun "dnf reinstall fapolicyd-$V_old-$R_old -y --allowerasing"
+        echo "allow perm=any all : all" >> /etc/fapolicyd/fapolicyd.rules
+        rlRun "ls -la /etc/fapolicyd/"
+        rlRun "dnf install fapolicyd-$V-$R -y --allowerasing"
+        rlRun "ls -la /etc/fapolicyd/"
+        rlRun "ls -la /etc/fapolicyd/rules.d/"
+        rlAssertExists /etc/fapolicyd/fapolicyd.rules
+        rlAssertEquals "rules are deployed into /etc/fapolicyd/rules.d" $(ls -1 /etc/fapolicyd/rules.d | wc -w) 0
+      fi
     rlPhaseEnd; }
 
     rlPhaseStartTest "upgrade to new version - still with fapolicyd.rules" && {
